requirements.txt
diff-match-patch==20121119
Django==2.0.5
django-cors-headers==2.2.0
django-import-export==1.0.1
djangorestframework==3.8.2
et-xmlfile==1.0.1
get==1.0.3
gunicorn==19.8.1
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.3
post==1.0.2
psycopg2==2.7.4
public==1.0.3
pytz==2018.4
PyYAML==3.12
query-string==1.0.2
request==1.0.2
tablib==0.12.1
unicodecsv==0.14.1
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
